klua working on pi3!
* Fixed superpage address when MMU disabled
* Bootcode now drops to EL1
* Enabled exception vectors
* Fixed some broken logic in the LDREX/STREX atomic fns
* Fixed zero padding for %lX printk format

+++ k/inc/std.h
@@ -41,7 +41,7 @@ typedef _Bool bool;
 // from a .word
 #define LOAD_WORD(reg, word) \
 	asm("MOVZ " #reg ", %0" : : "i" ((word) & 0xFFFF)); \
-	asm("MOVK " #reg ", %0, LSL #16" : : "i" (((uintptr)(word)) >> 16))
+	asm("MOVK " #reg ", %0, LSL #16" : : "i" (((uint32)(word)) >> 16))
 
 #else
 #define WORD(x) asm(".word %a0" : : "i" (x))
